backend/requirements.txt
pandas==2.0.3
python-dotenv==1.0.1
langchain
langchain_community
langchain_groq
langchain_core
langchain_huggingface
sentence_transformers
transformers
mlxtend==0.23.1
firebase-admin==6.0.1
google-cloud-storage==2.18.2
pinecone==5.3.1
motor==3.6.0
passlib==1.7.4
fastapi==0.115.4
uvicorn==0.32.0
pyjwt==2.9.0